feat(navigation): make the unified knowledge tree the primary Sidebar hierarchy (#461)
Make the unified knowledge tree the default hierarchy in the existing Sidebar, reuse one embedded tree on desktop and mobile, retire the standalone second drawer, retain a local legacy notebook-tree fallback, and add active-surface gating plus regression coverage.

Closes #460.
