+++ .github/workflows/knowledge-tree-sidebar-ci.yml
@@ -1,0 +1,65 @@
+name: Knowledge Tree Sidebar CI
+
+on:
+  pull_request:
+    branches: [main]
+    paths:
+      - 'frontend/src/components/Sidebar.tsx'
+      - 'frontend/src/components/KnowledgeTreeDrawer.tsx'
+      - 'frontend/src/components/KnowledgeTreePanel.tsx'
+      - 'frontend/src/lib/sidebarTreeMode.ts'
+      - 'frontend/src/lib/__tests__/sidebarTreeMode.test.ts'
+      - 'frontend/src/lib/__tests__/knowledgeTreeSidebarContract.test.ts'
+      - 'scripts/patch_issue460_*.py'
+      - '.github/workflows/knowledge-tree-sidebar-ci.yml'
+
+permissions:
+  contents: write
+
+jobs:
+  frontend:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Check out pull request branch
+        uses: actions/checkout@v4
+        with:
+          ref: ${{ github.event.pull_request.head.ref }}
+          fetch-depth: 0
+
+      - name: Apply controlled integration patches
+        shell: bash
+        run: |
+          shopt -s nullglob
+          patches=(scripts/patch_issue460_*.py)
+          if [ ${#patches[@]} -gt 0 ]; then
+            for patch in "${patches[@]}"; do
+              python3 "$patch"
+              rm "$patch"
+            done
+            git config user.name "github-actions[bot]"
+            git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
+            git add frontend/src/components/Sidebar.tsx frontend/src/components/KnowledgeTreePanel.tsx scripts
+            if ! git diff --cached --quiet; then
+              git commit -m "fix(navigation): keep one active knowledge tree surface"
+              git push origin HEAD:${{ github.event.pull_request.head.ref }}
+            fi
+          fi
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
+          cache: npm
+          cache-dependency-path: frontend/package-lock.json
+
+      - name: Install frontend dependencies
+        working-directory: frontend
+        run: npm ci --silent
+
+      - name: Sidebar mode and integration tests
+        working-directory: frontend
+        run: npx vitest run --reporter=basic src/lib/__tests__/sidebarTreeMode.test.ts src/lib/__tests__/knowledgeTreeSidebarContract.test.ts src/lib/__tests__/knowledgeTreeModel.test.ts
+
+      - name: Frontend typecheck
+        working-directory: frontend
+        run: npx tsc -b
